fix(validation): add .int() constraint to integer-typed fields
Added .int() to 10 fields that must be whole numbers: year fields
in leave and performance routes (with 2000-2100 range), year/month
in hs-metrics, headcount in departments, and sortOrder in objectives,
tasks, and fmea routes (3 schemas).

+++ apps/api-environment/src/routes/objectives.ts
@@ -256,7 +256,7 @@ router.patch('/:id/milestones/:milestoneId', async (req: AuthRequest, res: Respo
       title: z.string().optional(),
       dueDate: z.string().refine(s => !isNaN(Date.parse(s)), 'Invalid date format').optional(),
       notes: z.string().optional(),
-      sortOrder: z.number().nonnegative().optional(),
+      sortOrder: z.number().int().nonnegative().optional(),
     });
 
     const data = schema.parse(req.body);
